fix(handler): allow uint32 offset for migration scenarios (#7103)
(cherry picked from commit 5ce542b959a13242db9175383e69628333a5555c)

+++ internal/eventstore/handler/v2/statement.go
@@ -65,7 +65,7 @@ type Statement struct {
 	CreationDate  time.Time
 	InstanceID    string
 
-	offset uint16
+	offset uint32
 
 	Execute Exec
 }
